+++ Example/Example-watchOS Extension/InterfaceController.h
@@ -1,0 +1,18 @@
+//
+//  InterfaceController.h
+//  Example-watchOS Extension
+//
+//  Created by Jeff Kelley on 7/12/18.
+//  Copyright © 2018 Jeff Kelley. All rights reserved.
+//
+
+#import <WatchKit/WatchKit.h>
+#import <Foundation/Foundation.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface InterfaceController : WKInterfaceController
+
+@end
+
+NS_ASSUME_NONNULL_END